Refactored odds.fv
Also added some unit tests.  Some slight changes in the output format are noted in the test script.  The changes mostly involve treament of names.  No values changed except for a vector of NA's instead of a single NA for non-conforming input.  The refactoring simplifies the code greatly.  "mapply" is not needed because the individual conversion functions will accept a matrix.  The "get" function eliminates all the repetetive if-then's.